refactoring to make things more readable
